Replace start command with run, to avoid confusion with service lifetime commands.  Command still functions (for now) with a deprecation notice
