+++ releasenotes/notes/replcmds-3a95564a5f58197b.yaml
@@ -1,0 +1,10 @@
+---
+deprecations:
+  - |
+    Begin deprecating "Agent start".  It is being replaced by "run".  Command
+    will continue to function, with a deprecation notice
+other:
+  - |
+    Added new command "run" to the agent.  This command replacees the "start"
+    command, to reduce ambiguity with the service lifecycle commands
+
